feature: @UserId 어노테이션 추가
- @UserId 어노테이션 추가
- 연관된 Domain에 Serializable 추가
- 습관 단건 조회 시, userId로 user 비교하는 로직 추가
- 습관 삭제 시, userId로 user 비교하는 로직 추가
- 습관 생성 시, 습관에 해당하는 날짜가 없으면 예외 처리되도록 변경

+++ src/main/java/com/codeQuartette/myTime/service/impl/HabitServiceImpl.java
@@ -7,7 +7,10 @@ import com.codeQuartette.myTime.domain.HabitHasMyDate;
 import com.codeQuartette.myTime.domain.MyDate;
 import com.codeQuartette.myTime.domain.User;
 import com.codeQuartette.myTime.domain.value.Category;
+import com.codeQuartette.myTime.exception.CannotCreateHabitException;
 import com.codeQuartette.myTime.exception.HabitNotFoundException;
+import com.codeQuartette.myTime.exception.UserNotFoundException;
+import com.codeQuartette.myTime.exception.UserNotMatchException;
 import com.codeQuartette.myTime.repository.HabitRepository;
 import com.codeQuartette.myTime.service.HabitHasMyDateService;
 import com.codeQuartette.myTime.service.HabitService;
@@ -42,6 +45,9 @@ public class HabitServiceImpl implements HabitService {
         List<LocalDate> allHabitDates =
                 myDateService.checkAllDateByStartDateAndEndDate(habit.getStartDate(), habit.getEndDate(), habitRequestDTO.getRepeatDay());
         List<MyDate> newMyDates = allHabitDates.stream().map(date -> new MyDate(date, user)).toList();
+        if(newMyDates.size() == 0) {
+            throw new CannotCreateHabitException("습관에 해당하는 날짜가 없습니다. 다시 확인 후 습관을 생성해주세요.");
+        }
         List<MyDate> saveMyDates = myDateService.saveAllMyDate(newMyDates);
 
         habit = habitRepository.save(habit);
@@ -79,16 +85,26 @@ public class HabitServiceImpl implements HabitService {
 
     @Override
     @CacheEvict(cacheNames = "find-habit", allEntries = true)
-    public void delete(Long id) {
+    public void delete(Long userId, Long id) {
         Habit habit = habitRepository.findById(id)
                 .orElseThrow(() -> new HabitNotFoundException("삭제하려는 습관을 조회할 수 없습니다."));
+
+        if(!habit.getHabitHasMyDates().get(0).getMyDate().matchUser(userId)) {
+            throw new UserNotMatchException();
+        }
         habitRepository.delete(habit);
     }
 
     @Override
-    public Habit findHabit(Long id) {
-        return habitRepository.findById(id)
+    @Transactional
+    public Habit findHabit(Long userId, Long id) {
+        Habit habit = habitRepository.findById(id)
                 .orElseThrow(() -> new HabitNotFoundException("해당 습관을 조회할 수 없습니다."));
+
+        if(!habit.getHabitHasMyDates().get(0).getMyDate().matchUser(userId)) {
+            throw new UserNotMatchException();
+        }
+        return habit;
     }
 
     @Override
